Refactor : separate Main && parallax section background from content
