+++ src/components/molecules/ParallaxImages.tsx
@@ -1,0 +1,53 @@
+import { motion, useScroll, useTransform } from "motion/react";
+
+import { styled } from "styled-components";
+
+import { useRef } from "react";
+import { QUESTIONS } from "../../constants/textConstants";
+
+const ParallaxImages = ({ activeIndex }: { activeIndex: number }) => {
+  const containerRef = useRef<HTMLDivElement | null>(null);
+
+  const { scrollYProgress } = useScroll({
+    target: containerRef,
+    offset: ["start end", "end start"],
+  });
+
+  const y = useTransform(scrollYProgress, [0, 0.5, 1], ["-50%", "0%", "50%"]);
+
+  return (
+    <BackgroundWrapper ref={containerRef}>
+      {QUESTIONS.map(({ id }, i) => {
+        return (
+          <BackgroundImage
+            key={i}
+            srcSet={`/tennis/${id}-760.webp 760w, /tennis/${id}-1280.webp 1280w, /tennis/${id}-1920.webp 1920w, /tennis/${id}-2560.webp 2560w`}
+            style={{ y }}
+            animate={{
+              opacity: activeIndex === i ? 1 : 0,
+            }}
+            transition={{ duration: 0.6 }}
+            loading="lazy"
+          />
+        );
+      })}
+    </BackgroundWrapper>
+  );
+};
+
+export default ParallaxImages;
+
+const BackgroundWrapper = styled(motion.div)`
+  position: sticky;
+  top: 0;
+  z-index: 7;
+  height: 100vh;
+`;
+
+const BackgroundImage = styled(motion.img)`
+  width: 100%;
+  height: 100vh;
+  position: absolute;
+  filter: grayscale(100%) brightness(70%) blur(10px);
+  object-fit: cover;
+`;
