+++ src/components/organism/content/HeroContent/index.tsx
@@ -1,0 +1,92 @@
+import { motion } from "motion/react";
+import { useState } from "react";
+import styled from "styled-components";
+import { handleFadeAnimation } from "../../../../utils/validation";
+import { AnimationProgressTypes } from "../../../../pages/Main";
+import { BottomArrowIcon } from "../../../atoms/Icon/index";
+
+const WELCOMEINTRO = "hello World!";
+
+const HeroContent = ({ isInView }: { isInView: boolean }) => {
+  const [animationProgress, setAnimationProgress] =
+    useState<AnimationProgressTypes>("SCALE");
+
+  return (
+    <ContentWrapper
+      variants={contentOpacity}
+      initial="hidden"
+      animate={handleFadeAnimation({ state: animationProgress, isInView })}
+      onAnimationComplete={() =>
+        setAnimationProgress((prev) => (prev === "SCALE" ? "FADE" : prev))
+      }
+    >
+      {WELCOMEINTRO.toUpperCase()
+        .split(" ")
+        .map((text, i) => {
+          return (
+            <motion.p key={i} variants={contentTextItemReveal}>
+              {text}
+            </motion.p>
+          );
+        })}
+      {animationProgress === "FADE" && (
+        <IconWrapper
+          initial={{ y: 10, opacity: 0 }}
+          animate={{ y: [10, 0, 10], opacity: 1 }}
+          transition={{
+            duration: 1,
+            repeat: Infinity,
+            repeatType: "reverse",
+            type: "tween",
+          }}
+        >
+          <BottomArrowIcon />
+        </IconWrapper>
+      )}
+    </ContentWrapper>
+  );
+};
+
+export default HeroContent;
+
+const contentTextItemReveal = {
+  hidden: {
+    opacity: 0,
+    y: 10,
+  },
+  show: {
+    opacity: 1,
+    y: 0,
+  },
+};
+
+const IconWrapper = styled(motion.div)`
+  position: absolute;
+  width: 10vw;
+  aspect-ratio: 1/1;
+  bottom: -50%;
+`;
+
+const contentOpacity = {
+  hidden: { opacity: 0 },
+  show: {
+    opacity: 1,
+    transition: {
+      staggerChildren: 0.2,
+    },
+  },
+  exit: {
+    display: "none",
+  },
+};
+
+const ContentWrapper = styled(motion.h1)`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  z-index: 90;
+`;
