+++ src/components/template/section/SliderSection/index.tsx
@@ -7,22 +7,19 @@ import {
   useMotionTemplate,
   useMotionValueEvent,
 } from "framer-motion";
-
+import useScreenStore, { Mode } from "../../../../stores/useScreenStore";
 import {
-  BOOKSTITLE,
   BOOK_SECTION_HEIGHT,
+  BOOKSTITLE,
   CARD_TOTAL_WIDTH,
   CARD_WRAPPER_GAP,
   CARD_WRAPPER_WIDTH,
   INITIAL_Y_OFFSET,
   LAST_Y_OFFSET,
-} from "../../constants/booksConstants";
-
-import useScreenStore, { Mode } from "../../stores/useScreenStore";
-import { isScrollingBookSection } from "../../utils/calculation";
-import BookBackground from "../atoms/BookBackground";
-
-import BookCard from "../molecules/BookCard";
+} from "../../../../constants/booksConstants";
+import { isScrollingBookSection } from "../../../../utils/calculation";
+import BookBackground from "../../../atoms/BookBackground";
+import BookCard from "../../../molecules/BookCard";
 
 //background 의 setInAcitive 또한 카드의 넓이만큼 이동했을때 변해야한다 .
 // 한 화면에 카드를 두장씩 보여주기 위해선 하나의 카드가 100vw 만큼의 넓이를 가져가면 볼 수 없다.
@@ -153,6 +150,7 @@ const Container = styled.section<{
   height: ${({ $totalBooks, $bookSectionHeight }) =>
     $totalBooks * $bookSectionHeight}vh;
   background-color: black;
+  z-index: 6;
 `;
 
 const StickyArea = styled.div<{ $mode: Mode }>`
