+++ src/layout/Header.tsx
@@ -3,6 +3,7 @@ import OverlayHeader from "../components/molecules/OverlayHeader";
 import ExpandeHeader from "../components/molecules/ExpandeHeader";
 import useScreenMode from "../hooks/useScreenMode";
 import useScreenStore from "../stores/useScreenStore";
+import Stripes from "./background/Stripes";
 
 const Header = () => {
   useScreenMode();
@@ -11,7 +12,7 @@ const Header = () => {
   return (
     <>
       {mode !== "mobile" ? <ExpandeHeader /> : <OverlayHeader />}
-      {/* <Stripes /> */}
+      <Stripes />
       <Outlet />
     </>
   );
