+++ src/utils/calculation.ts
@@ -6,8 +6,6 @@ import {
 } from "../constants/skillsConstants";
 import { Mode } from "../stores/useScreenStore";
 
-import { MotionValue } from "motion";
-
 export const isScrollingBookSection = (n: number) => n > 0 && n < 1;
 
 export const getSkillIconWidth = (mode: Mode) => {
@@ -38,20 +36,4 @@ export const getScratchHighlightIndex = (
 export const getParallaxActiveIndex = (latest: number, dataLength: number) => {
   const newIndex = Math.min(Math.floor(latest * dataLength), dataLength - 1);
   return newIndex;
-};
-
-export const getImageParallaxYByOrder = ({
-  currentIdx,
-  totalLength,
-  initialY,
-  lastY,
-}: {
-  currentIdx: number;
-  totalLength: number;
-  initialY: MotionValue<string>;
-  lastY: MotionValue<string>;
-}) => {
-  const y =
-    currentIdx === 0 ? initialY : currentIdx === totalLength - 1 ? lastY : 0;
-  return y;
 };
